9_DeepLabV3_Image_Segmentation_with_Foreground_Background_Separation_and_Text_Addition/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cryptography==43.0.3
cycler==0.12.1
dataclasses==0.6
deepface==0.0.93
efficientnet==1.1.1
filelock==3.16.1
fire==0.7.0
flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.4.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.67.1
gTTS==2.5.4
gunicorn==23.0.0
h5py==3.11.0
idna==3.10
image-classifiers==1.0.0
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jax==0.4.13
jaxlib==0.4.13
jinja2==3.1.4
keras==2.13.1
Keras-Applications==1.0.8
kiwisolver==1.4.7
lazy-loader==0.4
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
mpmath==1.3.0
mtcnn==0.1.1
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.4.0
packaging==24.1
pandas==2.0.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
protobuf==3.20.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydub==0.25.1
pygame==2.6.1
PyMuPDF==1.24.11
pyparsing==3.1.4
pypdf2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
segmentation-models==1.0.1
six==1.16.0
smmap==5.0.1
smplx==0.1.28
sounddevice==0.5.1
soupsieve==2.6
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf-keras==2.15.1
tifffile==2023.7.10
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.0
triton==3.0.0
typing-extensions==4.5.0
tzdata==2024.2
ultralytics==8.3.31
ultralytics-thop==2.0.11
urllib3==2.2.3
werkzeug==3.0.6
wrapt==1.16.0
zipp==3.20.2
